Finalize lazy functions
